Added Protection trigger for uci push actions, it's still getting to guege reset! Some times, not any more!

+++ sbin/ext/post-init.sh
@@ -116,17 +116,25 @@ echo "1" > /sys/devices/platform/samsung-pd.2/mdnie/mdnie/mdnie/user_mode
 /sbin/busybox chmod 755 /res/customconfig/actions/push-actions/*
 /sbin/busybox mv /res/customconfig/actions/push-actions/* /res/no-push-on-boot/
 
-
+(
 # apply ExTweaks defaults
 # in case we forgot to set permissions, fix them.
 /sbin/busybox chmod 755 /res/customconfig/actions/*
 /res/uci.sh apply
+echo "uci done" > /data/.siyah/uci_loaded
+)&
 
 (
-sleep 40
-# Restore all the PUSH Button Actions back to there location.
-/sbin/busybox mount -o remount,rw rootfs
-/sbin/busybox mv /res/no-push-on-boot/* /res/customconfig/actions/push-actions/
+while : ; do
+	if [ -e /data/.siyah/uci_loaded ] ; then
+		# Restore all the PUSH Button Actions back to there location.
+		/sbin/busybox mount -o remount,rw rootfs
+		/sbin/busybox mv /res/no-push-on-boot/* /res/customconfig/actions/push-actions/
+		rm -f /data/.siyah/uci_loaded
+		exit 0
+	fi;
+	sleep 5
+done
 )&
 
 ##### init scripts #####
